combines constants and fasttrig into dsp/maths
