+++ src/svf.c
@@ -1,5 +1,5 @@
 
-#include <dsp/constants.h>
+#include <dsp/maths.h>
 #include <dsp/shape.h>
 #include <dsp/svf.h>
 #include <dsp/utils.h>
